Feature ssh key pair

+++ dcmgr/lib/dcmgr/endpoints/12.03/ssh_key_pairs.rb
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 require 'dcmgr/endpoints/12.03/responses/ssh_key_pair'
-
+require 'sshkey'
 Dcmgr::Endpoints::V1203::CoreAPI.namespace '/ssh_key_pairs' do
   get do
     ds = M::SshKeyPair.dataset
@@ -21,6 +21,8 @@ Dcmgr::Endpoints::V1203::CoreAPI.namespace '/ssh_key_pairs' do
       ds = ds.filter(:display_name=>params[:display_name])
     end
 
+    ds = ds.alives
+
     collection_respond_with(ds) do |paging_ds|
       R::SshKeyPairCollection.new(paging_ds).generate
     end
@@ -39,18 +41,26 @@ Dcmgr::Endpoints::V1203::CoreAPI.namespace '/ssh_key_pairs' do
   quota 'ssh_key_pair.count'
   post do
     # description "Create ssh key pair information"
-    # params :download_once optional set true if you do not want
-    #        to save private key info on database.
     # params :display_name optional
-    keydata = nil
-
+    private_key = nil
     ssh = M::SshKeyPair.entry_new(@account) do |s|
-      keydata = M::SshKeyPair.generate_key_pair(s.uuid)
-      s.public_key = keydata[:public_key]
-      s.finger_print = keydata[:finger_print]
 
-      if params[:download_once] != 'true'
-        s.private_key = keydata[:private_key]
+      unless params[:public_key].empty?
+        public_key = URI.decode(params[:public_key])
+        error = 0
+        error += 1 unless SSHKey.valid_ssh_public_key?(public_key)
+        result = `/usr/bin/ssh-keygen -lf /dev/stdin <<< '#{public_key}'`
+        error += 1 unless $? == 0
+        raise InvalidSshPublicKey, params[:public_key] if  error > 0
+
+        s.public_key = public_key
+        s.finger_print = result.split(' ')[1]
+      else
+        keydata = nil
+        keydata = M::SshKeyPair.generate_key_pair(s.uuid)
+        private_key = keydata[:private_key]
+        s.public_key = keydata[:public_key]
+        s.finger_print = keydata[:finger_print]
       end
 
       if params[:description]
@@ -73,14 +83,26 @@ Dcmgr::Endpoints::V1203::CoreAPI.namespace '/ssh_key_pairs' do
       raise E::DatabaseError, e.message
     end
 
-    respond_with(R::SshKeyPair.new(ssh, keydata[:private_key]).generate)
+    respond_with(R::SshKeyPair.new(ssh, private_key).generate)
   end
 
   delete '/:id' do
     # description "Remove ssh key pair information"
     # params :id required
     ssh = find_by_uuid(:SshKeyPair, params[:id])
-    ssh.destroy
+    raise E::UnknownSshKeyPair, params[:id] if ssh.nil?
+
+    force = false
+    if params[:force] == 'true'
+      force = true
+    end
+
+    begin
+      ssh.force = force
+      ssh.destroy
+    rescue => e
+      raise E::ExistsRegisteredInstance, e.message
+    end
 
     respond_with([ssh.canonical_uuid])
   end
